Consolidate downstream handlers into acms-cams-transition
Replace the two separate handler packages (staff-assignment-handler,
trustee-appointment-handler) and the shared/ library with a single
acms-cams-transition module. Both Azure Function queue handlers, the
CmmapCamsRow type, and all shared helpers now live in one file,
reducing cognitive load and encapsulating the full ACMS integration
domain in one place. Also adds resolve.alias to the vitest config
so @common/* path aliases resolve correctly without mock factories.

Jira ticket: CAMS-616
