+++ backend/lib/use-cases/dataflows/sync-trustee-appointments.ts
@@ -161,7 +161,6 @@ async function applyResolvedTrustee(
         caseId: event.caseId,
         trusteeId: existingAppointment.trusteeId,
         acmsProfessionalId,
-        apptType: 'TR',
         assignedOn: existingAppointment.assignedOn,
         appointedDate: existingAppointment.appointedDate,
         chapter: syncedCase.chapter,
@@ -187,7 +186,6 @@ async function applyResolvedTrustee(
       caseId: event.caseId,
       trusteeId,
       acmsProfessionalId,
-      apptType: 'TR',
       assignedOn: now,
       appointedDate: event.appointedDate,
       chapter: syncedCase.chapter,
