+++ downstream/index.ts
@@ -1,2 +1,1 @@
-import './staff-assignment-handler/staff-assignment-handler';
-import './trustee-appointment-handler/trustee-appointment-handler';
+import './acms-cams-transition/acms-cams-transition';
