+++ downstream/vitest.config.mjs
@@ -1,6 +1,15 @@
 import { defineConfig } from 'vitest/config';
+import path from 'path';
+import { fileURLToPath } from 'url';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 export default defineConfig({
+  resolve: {
+    alias: {
+      '@common': path.resolve(__dirname, '../common/src'),
+    },
+  },
   test: {
     globals: true,
     environment: 'node',
